# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warehouse Management System Stub API
  version: 1.0.0
  description: |
    Versioned backend scaffold for web and mobile clients.
    All endpoints are stubs returning mock envelopes.
servers:
  - url: http://localhost:5000/api/v1
    description: Local backend
security:
  - bearerAuth: []
paths:
...
  /putaway/tasks:
    get:
      tags: [Putaway]
      summary: List putaway tasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assignee
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/ListSuccessResponse'
...
  /picking/tasks:
    get:
      tags: [Picking]
      summary: List picking tasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assignee
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/ListSuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 500
  responses:
...
    ListSuccessResponse:
      description: Success envelope with pagination metadata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListSuccessEnvelope'
...
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
        tenantId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required: [requestId, tenantId, timestamp]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, total, totalPages]
...
    ListSuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/Pagination'
      required: [success, data, meta]
...